AWS Config File Checks
Refactored and tested version
