--- myuser/myuser.go
@@ -1,16 +1,0 @@
-package myuser
-
-import (
-	"os/user"
-)
-
-// GetUserHomeDir gets path to home directory.
-func GetUserHomeDir() (string, error) {
-	myUser, userError := user.Current()
-	if userError != nil {
-		return "", userError
-	}
-	path := myUser.HomeDir
-
-	return path, nil
-}
